+++ src/bin/main.rs
@@ -162,9 +162,11 @@ fn main() {
                                         .unwrap()
                                         .map(|s| s.as_str())
                                         .collect::<Vec<&str>>(),
+                                        cli_subargs.get_flag("regex"),
                                     cli_subargs.get_flag("skip"),
                                     cli_subargs.get_flag("count"),
                                     cli_subargs.get_flag("force"),
+                                    cli_subargs.get_one::<usize>("sub").copied(),
                                     *cli_subargs.get_one::<u64>("seed").unwrap(),
                                     &logger,
                                     *cli_subargs.get_one::<usize>("threads").unwrap(),
@@ -191,6 +193,7 @@ fn main() {
                                         .unwrap()
                                         .map(|s| s.as_str())
                                         .collect::<Vec<&str>>(),
+                                    cli_subargs.get_flag("regex"),
                                         cli_subargs
                                         .get_many::<String>("lang")
                                         .map(|v|